--- a/packages/block/src/header.ts
+++ b/packages/block/src/header.ts
@@ -65,6 +65,9 @@ export class BlockHeader {
 
   public readonly common: Common
 
+  private keccakFunction?: Function
+  private static keccakFunction: ((msg: Uint8Array) => Uint8Array) | undefined
+
   private cache: HeaderCache = {
     hash: undefined,
   }
@@ -89,6 +92,7 @@ export class BlockHeader {
    * @param opts
    */
   public static fromHeaderData(headerData: HeaderData = {}, opts: BlockOptions = {}) {
+    this.keccakFunction = this.keccakFunction ?? opts?.common?.customCrypto.keccak256 ?? keccak256
     return new BlockHeader(headerData, opts)
   }
 
@@ -99,6 +103,7 @@ export class BlockHeader {
    * @param opts
    */
   public static fromRLPSerializedHeader(serializedHeaderData: Uint8Array, opts: BlockOptions = {}) {
+    this.keccakFunction = this.keccakFunction ?? opts?.common?.customCrypto.keccak256 ?? keccak256
     const values = RLP.decode(serializedHeaderData)
     if (!Array.isArray(values)) {
       throw new Error('Invalid serialized header input. Must be array')
@@ -113,6 +118,7 @@ export class BlockHeader {
    * @param opts
    */
   public static fromValuesArray(values: BlockHeaderBytes, opts: BlockOptions = {}) {
+    this.keccakFunction = this.keccakFunction ?? opts?.common?.customCrypto.keccak256 ?? keccak256
     const headerData = valuesArrayToHeaderData(values)
     const { number, baseFeePerGas, excessBlobGas, blobGasUsed, parentBeaconBlockRoot } = headerData
     const header = BlockHeader.fromHeaderData(headerData, opts)
@@ -152,6 +158,8 @@ export class BlockHeader {
         chain: Chain.Mainnet, // default
       })
     }
+    this.keccakFunction =
+      this.keccakFunction ?? options?.common?.customCrypto.keccak256 ?? keccak256
 
     const skipValidateConsensusFormat = options.skipConsensusFormatValidation ?? false
 
@@ -672,11 +680,11 @@ export class BlockHeader {
   hash(): Uint8Array {
     if (Object.isFrozen(this)) {
       if (!this.cache.hash) {
-        this.cache.hash = keccak256(RLP.encode(this.raw()))
+        this.cache.hash = this.keccakFunction!(RLP.encode(this.raw()))
       }
       return this.cache.hash
     }
-    return keccak256(RLP.encode(this.raw()))
+    return this.keccakFunction!(RLP.encode(this.raw()))
   }
 
   /**
@@ -775,7 +783,7 @@ export class BlockHeader {
     this._requireClique('cliqueSigHash')
     const raw = this.raw()
     raw[12] = this.extraData.subarray(0, this.extraData.length - CLIQUE_EXTRA_SEAL)
-    return keccak256(RLP.encode(raw))
+    return this.keccakFunction!(RLP.encode(raw))
   }
 
   /**