--- a/packages/client/bin/cli.ts
+++ b/packages/client/bin/cli.ts
@@ -7,14 +7,25 @@ import { RLP } from '@ethereumjs/rlp'
 import {
   Address,
   bytesToHex,
+  calculateSigRecovery,
+  concatBytes,
+  ecrecover,
   hexToBytes,
   initKZG,
   parseGethGenesisState,
   randomBytes,
+  setLengthLeft,
   short,
   toBytes,
 } from '@ethereumjs/util'
+import {
+  keccak256 as keccak256WASM,
+  secp256k1Expand,
+  secp256k1Recover,
+  waitReady,
+} from '@polkadot/wasm-crypto'
 import * as kzg from 'c-kzg'
+import { keccak256 } from 'ethereum-cryptography/keccak'
 import { existsSync, writeFileSync } from 'fs'
 import { ensureDirSync, readFileSync, removeSync } from 'fs-extra'
 import { Level } from 'level'
@@ -38,6 +49,7 @@ import type { FullEthereumService } from '../src/service'
 import type { ClientOpts } from '../src/types'
 import type { RPCArgs } from './startRpc'
 import type { BlockBytes } from '@ethereumjs/block'
+import type { CustomCrypto } from '@ethereumjs/common'
 import type { GenesisState } from '@ethereumjs/util'
 import type { AbstractLevel } from 'abstract-level'
 
@@ -401,6 +413,11 @@ const args: ClientOpts = yargs(hideBin(process.argv))
     describe:
       'Skip executing blocks in new payload calls in engine, alias for --engineNewpayloadMaxExecute=0 and overrides any engineNewpayloadMaxExecute if also provided',
     boolean: true,
+  })
+  .option('useJsCrypto', {
+    describe: 'Use pure Javascript cryptography functions',
+    boolean: true,
+    default: false,
   })
   .completion()
   // strict() ensures that yargs throws when an invalid arg is provided
@@ -782,7 +799,28 @@ async function run() {
   initKZG(kzg, args.trustedSetup ?? __dirname + '/../src/trustedSetups/official.txt')
   // Give network id precedence over network name
   const chain = args.networkId ?? args.network ?? Chain.Mainnet
-
+  const cryptoFunctions: CustomCrypto = {}
+  if (args.useJsCrypto === false) {
+    await waitReady()
+    cryptoFunctions.keccak256 = keccak256WASM
+    cryptoFunctions.ecrecover = (
+      msgHash: Uint8Array,
+      v: bigint,
+      r: Uint8Array,
+      s: Uint8Array,
+      chainID?: bigint
+    ) =>
+      secp256k1Expand(
+        secp256k1Recover(
+          msgHash,
+          concatBytes(setLengthLeft(r, 32), setLengthLeft(s, 32)),
+          Number(calculateSigRecovery(v, chainID))
+        )
+      ).slice(1)
+  } else {
+    cryptoFunctions.keccak256 = keccak256
+    cryptoFunctions.ecrecover = ecrecover
+  }
   // Configure accounts for mining and prefunding in a local devnet
   const accounts: Account[] = []
   if (typeof args.unlock === 'string') {
@@ -790,7 +828,7 @@ async function run() {
   }
 
   let customGenesisState: GenesisState | undefined
-  let common = new Common({ chain, hardfork: Hardfork.Chainstart })
+  let common = new Common({ chain, hardfork: Hardfork.Chainstart, customCrypto: cryptoFunctions })
 
   if (args.dev === true || typeof args.dev === 'string') {
     args.discDns = false
@@ -814,6 +852,7 @@ async function run() {
       common = new Common({
         chain: customChainParams.name,
         customChains: [customChainParams],
+        customCrypto: cryptoFunctions,
       })
     } catch (err: any) {
       console.error(`invalid chain parameters: ${err.message}`)
@@ -827,6 +866,8 @@ async function run() {
       chain: chainName,
       mergeForkIdPostMerge: args.mergeForkIdPostMerge,
     })
+    //@ts-ignore
+    common.customCrypto = cryptoFunctions
     customGenesisState = parseGethGenesisState(genesisFile)
   }