… .github/workflows/client-build.yml
… packages/block/test/block.spec.ts
… packages/block/test/eip4895block.spec.ts
… packages/client/package.json
… packages/client/src/net/server/rlpxserver.ts
… packages/client/src/sync/fetcher/accountfetcher.ts
… packages/client/src/sync/fetcher/bytecodefetcher.ts
… packages/client/src/sync/fetcher/storagefetcher.ts
… packages/client/src/sync/fetcher/trienodefetcher.ts
… packages/client/src/types.ts
… packages/client/test/rpc/engine/withdrawals.spec.ts
… packages/client/test/util/wasmCrypto.spec.ts
… packages/client/vitest.config.unit.ts
… packages/common/src/common.ts
… packages/common/src/types.ts
… packages/common/test/customCrypto.spec.ts
… packages/devp2p/src/dns/dns.ts
… packages/devp2p/src/dns/enr.ts
… packages/devp2p/src/dpt/dpt.ts
… packages/devp2p/src/dpt/message.ts
… packages/devp2p/src/dpt/server.ts
… packages/devp2p/src/rlpx/ecies.ts
… packages/devp2p/src/rlpx/peer.ts
… packages/devp2p/src/rlpx/rlpx.ts
… packages/devp2p/src/types.ts
… packages/devp2p/src/util.ts
… packages/evm/src/opcodes/functions.ts
… packages/evm/src/opcodes/util.ts
… packages/evm/src/precompiles/01-ecrecover.ts
… packages/evm/src/precompiles/02-sha256.ts