This was actually the original triggger for #3187, hehe. 😁

So this PR tests the [@polkadot/wasm-crypto](https://github.com/polkadot-js/wasm/tree/master/packages/wasm-crypto) library as a replacement for the trie key hashing for the VM state manager in the client. Had this long on my plate to analyse this a bit better, since theoretically (re-)using a more native hash function should have a significant performance imact.

So I tested with the following early mainnet block (now starting directly with the block to avoid state caches):

```shell
npm run client:start -- --vmProfileBlocks --executeBlocks=227715
```

Block has 79 simple value transfer txs, so not much else is happening and mainly the sender and receiver state changes are written and come into play here.

Following results (with just the first two commits, so only exchanging for trie in execution):

Before (Entire Block):  280ms  277ms  276ms  283ms. 285ms  283ms. 280ms. 279ms
After.   (Entire Block):  264ms. 270ms. 265ms. 268ms. 266ms. 262ms. 263ms. 265ms

Results are significant enough that I would continue here, especially since these are results for *only* exchanging things for the execution state trie, I expect this to get even substantially further if we do this holistically and exchange for all other places where keccak256 hashing is applied (tx trie validations, keccak256 precompile + opcode (there are these two, right?), guess some other things as well).

I will work a bit on this since I would like to get this done in a clean way. If I am not mistaken our whole "replace the hash function" functionality/API is generally not complete (so pretty sure e.g. for VM this is neither possible in the precompile or anywhere + things (the changed hash function) would at the end also needed to passed down to the lower level libraries.