@acolytec3 2023-12-18
I'm liking this new approach. I wonder if we should do the same with KZG and move away from the global singleton