@holgerd77 2023-12-29
Just compiling a list here on what is still missing, in case its helpful (no matter who continues on this):

- [ ] Client: pass `common` to various trie instantiations, particularly the new Snap fetchters, search for `new Trie(` (no additional `Trie.create(` instance apart from VM execution (already updated))
- [ ] Client: bytecode/trienode fetcher hashing (if easily doable, so in `sync/fetcher`)
- [ ] Block: pass `common` to withdrawals and tx trie (this static `genTransactionsTrieRoot()` needs a closer look here)
- [ ] Block: use keccak for hashing (block/uncle hash)
- [ ] StateManager: pass `common` to internal `_trie` object
- [ ] StateManager: `keccak256` usages (various)
- [ ] VM: receipt trie (in `runBlock()` and `buildBlock()`), also `_genTxTrie()` in `runBlock()`
- [ ] EVM: `AUTH` opcode (optional but very easy)
- [ ] Trie: pass to new instantiation in `verifyProof()`
- [ ] Trie: I think `shallowCopy()` does not work yet, might need confirmation
- [ ] Devp2p: various `keccak256` usages (all over the place 🤯, totally wasn't aware actually)
- [ ] All touched: generall give `shallowCopy()` a quick look (eventually, optimally: a test 😬)

And finally, tricky one but I guess important: `Account` in the Util package. Ugh. Likely again with an interface as in trie and then apply where used (mainly EVM?).


Ugh. This list went longer than I expected. 😆 🙂

But on the other hand: if WASM is really that much faster this also REALLY will have an impact if applied so broadly!