@holgerd77 2024-01-11
Update: I think I will just work a bit on this myself, maybe I will get a better grasp why this static/non-static thing was done or if it was just an overlook.

Will push my work (if something useful comes out).