@holgerd77 2024-01-11
(and update on the tests: ah, ok, this is "only" this package resolution thing, then discard my non-continue message from above)