@holgerd77 2024-01-11
Gave devp2p another look.

This might be rather a thing for a new PR. 

This seems to need some various structural changes and this would likely overload the PR here.